app/requirements.txt
flask
werkzeug
requests
huggingface_hub
flask-session
pyjwt